Test_work/requirements.txt
#Python==3.9.1
asgiref==3.5.0
attrs==21.4.0
bcrypt==3.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
colorama==0.4.4
cryptography==36.0.1
distro==1.7.0
Django==3.2.5
django-filter==21.1
djangorestframework==3.13.1
djongo==1.3.6
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
idna==3.3
jsonschema==3.2.0
mongoengine==0.24.0
paramiko==2.9.2
pycparser==2.21
pymongo==3.12.1
PyNaCl==1.5.0
pyrsistent==0.18.1
python-dotenv==0.19.2
pytz==2021.3
#pywin32==227
PyYAML==5.4.1
requests==2.27.1
six==1.16.0
sqlparse==0.2.4
texttable==1.6.4
tzdata==2021.5
urllib3==1.26.8
websocket-client==0.59.0
